+++ examples/master_querant.rs
@@ -1,9 +1,10 @@
 
 use std::env;
-use gamedig::{aliens, aoc, arma2oa, ase, asrd, avorion, bat1944, bb2, bm, bo, ccure, cosu, cs, cscz, csgo, css, dod, dods, doi, dst, GDResult, gm, hl2dm, hldms, ins, insmic, inss, l4d, l4d2, mc, ohd, onset, pz, ror2, rust, sc, sdtd, tf, tf2, tfc, ts, unturned, vr};
+use gamedig::{aliens, aoc, arma2oa, ase, asrd, avorion, bat1944, bb2, bf1942, bm, bo, ccure, cosu, cs, cscz, csgo, css, dod, dods, doi, dst, GDResult, gm, hl2dm, hldms, ins, insmic, inss, l4d, l4d2, mc, ohd, onset, pz, ror2, rust, sc, sdtd, tf, tf2, tfc, ts, unturned, ut, vr};
 use gamedig::protocols::minecraft::LegacyGroup;
 use gamedig::protocols::valve;
 use gamedig::protocols::valve::Engine;
+use gamedig::protocols::gamespy;
 
 fn main() -> GDResult<()> {
     let args: Vec<String> = env::args().collect();
@@ -83,6 +84,10 @@ fn main() -> GDResult<()> {
         "avorion" => println!("{:#?}", avorion::query(ip, port)?),
         "ohd" => println!("{:#?}", ohd::query(ip, port)?),
         "vr" => println!("{:#?}", vr::query(ip, port)?),
+        "_gamespy1" => println!("{:#?}", gamespy::one::query(ip, port.unwrap(), None)),
+        "_gamespy1_vars" => println!("{:#?}", gamespy::one::query_vars(ip, port.unwrap(), None)),
+        "ut" => println!("{:#?}", ut::query(ip, port)),
+        "bf1942" => println!("{:#?}", bf1942::query(ip, port)),
         _ => panic!("Undefined game: {}", args[1])
     };