+++ src/games/bf1942.rs
@@ -1,0 +1,10 @@
+use crate::GDResult;
+use crate::protocols::gamespy;
+use crate::protocols::gamespy::Response;
+
+pub fn query(address: &str, port: Option<u16>) -> GDResult<Response> {
+    gamespy::one::query(address, match port {
+        None => 23000,
+        Some(port) => port
+    }, None)
+}