+++ src/protocols/gamespy/protocol/mod.rs
@@ -1,0 +1,3 @@
+
+/// GameSpy 1
+pub mod one;