+++ src/protocols/gamespy/types.rs
@@ -1,0 +1,36 @@
+use std::collections::HashMap;
+
+/// A player’s details.
+#[derive(Debug)]
+pub struct Player {
+    pub name: String,
+    pub team: u8,
+    /// The ping from the server's perspective.
+    pub ping: u16,
+    pub face: String,
+    pub skin: String,
+    pub mesh: String,
+    pub frags: u32,
+    pub deaths: Option<u32>,
+    pub health: Option<u32>,
+    pub secret: bool
+}
+
+/// A query response.
+#[derive(Debug)]
+pub struct Response {
+    pub name: String,
+    pub map: String,
+    pub map_title: Option<String>,
+    pub admin_contact: Option<String>,
+    pub admin_name: Option<String>,
+    pub has_password: bool,
+    pub game_type: String,
+    pub game_version: String,
+    pub players_maximum: usize,
+    pub players_online: usize,
+    pub players_minimum: u8,
+    pub players: Vec<Player>,
+    pub tournament: bool,
+    pub unused_entries: HashMap<String, String>
+}